feat(class): prettier or rename import (due to rename files)
