+++ src/app/[locale]/flashcards/learning/page.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useRef, useState } from 'react';
 import useFetch from '@/hooks/useFetch';
 import { putRequest } from '@/api/api';
-import { IFlashcardFull, IQualityResponse, IQualityResponseNextReviewInterval } from '../flashcard.type';
+import { IFlashcardFull, IQualityResponse, IQualityResponseNextReviewInterval } from '../types/flashcard.type';
 import LoadingPage from '@/app/loading';
 import { FlashcardLearning } from './components/FlashcardLearning';
